earlier, stable release of circuitpython (10.0.3)

+++ ports/stm/peripherals/timers.h
@@ -7,12 +7,13 @@
 #pragma once
 
 #include <stdint.h>
-#include <stdbool.h>
+#include "py/mphal.h"
+#include "peripherals/periph.h"
 
 #include STM32_HAL_H
 
-void tim_clock_enable(uint32_t mask);
-void tim_clock_disable(uint32_t mask);
+void tim_clock_enable(uint16_t mask);
+void tim_clock_disable(uint16_t mask);
 uint32_t stm_peripherals_timer_get_source_freq(TIM_TypeDef *timer);
 size_t stm_peripherals_timer_get_irqnum(TIM_TypeDef *instance);
 void timers_reset(void);
